+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## [UNRELEASED]
+### Added
+- Send unscheduled alarm event
+
 ## [1.0.0] - 2022-01-15
 
 ### Added